gateway-api  refactor
