+++ gateway-api/knotx/conf/routes/handlers/paymentRxHandlerWithCB.conf
@@ -15,11 +15,13 @@ paypal {
     resetTimeout = 10000
     timeout = 200
   }
+
   fallback {
     verified = true
     paymentUrl = "https://paypal-fallback.com/payment?id=1983247919hv9sa398f"
     label = "PayPal standard"
   }
+
   host = "webapi"
   port = 8080
   requestUri = "/paypal/verify"
