+++ gateway-api/src/main/java/io/knotx/example/payment/action/PaymentsActionFactory.java
@@ -1,9 +1,11 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package io.knotx.example.action.payments;
+package io.knotx.example.payment.action;
 
-import static io.knotx.example.action.payments.ProvidersProvider.calculateProviders;
+import static io.knotx.example.payment.utils.ProvidersProvider.calculateProviders;
+
+import org.apache.commons.lang3.StringUtils;
 
 import io.knotx.fragments.handler.api.Action;
 import io.knotx.fragments.handler.api.ActionFactory;
@@ -12,7 +14,6 @@ import io.reactivex.Single;
 import io.vertx.core.Future;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
-import org.apache.commons.lang3.StringUtils;
 
 public class PaymentsActionFactory implements ActionFactory {
