+++ gateway-api/src/main/java/io/knotx/example/payment/handler/rx/PaymentRxHandlerFactory.java
@@ -1,4 +1,4 @@
-package io.knotx.example.api.payment;
+package io.knotx.example.payment.handler.rx;
 
 import io.knotx.server.api.handler.RoutingHandlerFactory;
 import io.vertx.core.Handler;
@@ -6,16 +6,16 @@ import io.vertx.core.json.JsonObject;
 import io.vertx.reactivex.core.Vertx;
 import io.vertx.reactivex.ext.web.RoutingContext;
 
-public class PaymentRxHandlerWithCircuitBreakerFactory implements RoutingHandlerFactory {
+public class PaymentRxHandlerFactory implements RoutingHandlerFactory {
 
   @Override
   public String getName() {
-    return "paymentRxHandlerWithCircuitBreaker";
+    return "paymentRxHandler";
   }
 
   @Override
   public Handler<RoutingContext> create(Vertx vertx, JsonObject config) {
-    return new PaymentRxHandlerWithCircuitBreaker(vertx, config);
+    return new PaymentRxHandler(vertx, config);
   }
 
 }
