+++ gateway-api/src/main/resources/META-INF/services/io.knotx.fragments.handler.api.ActionFactory
@@ -1,0 +1,1 @@
+io.knotx.example.payment.action.PaymentsActionFactory
